+++ how-to/workspace-platform-starter/.eslintrc.extended.js
@@ -20,6 +20,46 @@ module.exports = {
 					{
 						selector: "BinaryExpression[operator='!=='][right.type='Identifier'][right.name='null']",
 						message: 'Instead of using "!== null", please use the utils method !isEmpty'
+					},
+					{
+						selector:
+							"BinaryExpression[operator='==='][left.type='UnaryExpression'][left.operator='typeof'][right.type='Literal'][right.value='object']",
+						message: 'Instead of using "=== object", please use the utils method isObject'
+					},
+					{
+						selector:
+							"BinaryExpression[operator='==='][left.type='UnaryExpression'][left.operator='typeof'][right.type='Literal'][right.value='object']",
+						message: 'Instead of using "!== object", please use the utils method !isObject'
+					},
+					{
+						selector:
+							"BinaryExpression[operator='==='][left.type='UnaryExpression'][left.operator='typeof'][right.type='Literal'][right.value='string']",
+						message: 'Instead of using "=== string", please use the utils method isString'
+					},
+					{
+						selector:
+							"BinaryExpression[operator='==='][left.type='UnaryExpression'][left.operator='typeof'][right.type='Literal'][right.value='string']",
+						message: 'Instead of using "!== string", please use the utils method !isString'
+					},
+					{
+						selector:
+							"BinaryExpression[operator='==='][left.type='UnaryExpression'][left.operator='typeof'][right.type='Literal'][right.value='number']",
+						message: 'Instead of using "=== number", please use the utils method isNumber'
+					},
+					{
+						selector:
+							"BinaryExpression[operator='==='][left.type='UnaryExpression'][left.operator='typeof'][right.type='Literal'][right.value='number']",
+						message: 'Instead of using "!== number", please use the utils method !isNumber'
+					},
+					{
+						selector:
+							"BinaryExpression[operator='==='][left.type='UnaryExpression'][left.operator='typeof'][right.type='Literal'][right.value='boolean']",
+						message: 'Instead of using "=== boolean", please use the utils method isBoolean'
+					},
+					{
+						selector:
+							"BinaryExpression[operator='==='][left.type='UnaryExpression'][left.operator='typeof'][right.type='Literal'][right.value='boolean']",
+						message: 'Instead of using "!== boolean", please use the utils method !isBoolean'
 					}
 				]
 			}