+++ how-to/workspace-platform-starter/client/src/modules/integrations/pages/integration.ts
@@ -325,9 +325,13 @@ export class PagesProvider implements IntegrationModule<PagesSettings> {
 							// Deleting the page will eventually trigger the "delete" lifecycle
 							// event which will remove it from the result list
 						}
-					} else if (result.action.name === PagesProvider._ACTION_SHARE_PAGE) {
-						if (this._integrationHelpers?.share) {
-							await this._integrationHelpers.share({ type: "page", pageId: data.pageId });
+					} else if (
+						result.action.name === PagesProvider._ACTION_SHARE_PAGE &&
+						this._integrationHelpers?.getShareClient
+					) {
+						const shareClient = await this._integrationHelpers.getShareClient();
+						if (shareClient) {
+							await shareClient.share("page", { pageId: data.pageId });
 						}
 					} else {
 						handled = false;
@@ -483,10 +487,10 @@ export class PagesProvider implements IntegrationModule<PagesSettings> {
 
 		if (this._integrationHelpers && Array.isArray(pages)) {
 			let shareEnabled: boolean = false;
-			if (this._integrationHelpers?.getConditionsClient) {
-				const conditionsClient = await this._integrationHelpers.getConditionsClient();
-				if (conditionsClient) {
-					shareEnabled = await conditionsClient.check("sharing");
+			if (this._integrationHelpers?.getShareClient) {
+				const shareClient = await this._integrationHelpers.getShareClient();
+				if (shareClient) {
+					shareEnabled = await shareClient.typeEnabled("page");
 				}
 			}
 
@@ -575,10 +579,10 @@ export class PagesProvider implements IntegrationModule<PagesSettings> {
 
 				if (!isEmpty(lastPage)) {
 					let shareEnabled: boolean = false;
-					if (this._integrationHelpers?.getConditionsClient) {
-						const conditionsClient = await this._integrationHelpers.getConditionsClient();
-						if (conditionsClient) {
-							shareEnabled = await conditionsClient.check("sharing");
+					if (this._integrationHelpers?.getShareClient) {
+						const shareClient = await this._integrationHelpers.getShareClient();
+						if (shareClient) {
+							shareEnabled = await shareClient.typeEnabled("page");
 						}
 					}