+++ how-to/workspace-platform-starter/templates/webpack.config.js
@@ -312,6 +312,34 @@ const configs = [
 		experiments: {
 			outputModule: true
 		}
+	},
+	{
+		entry: './templates/src/share/index.ts',
+		devtool: 'inline-source-map',
+		module: {
+			rules: [
+				{
+					test: /\.tsx?$/,
+					use: 'ts-loader',
+					exclude: /node_modules/
+				}
+			]
+		},
+		resolve: {
+			extensions: ['.tsx', '.ts', '.js'],
+			alias
+		},
+		externals: { fin: 'fin' },
+		output: {
+			filename: 'share.bundle.js',
+			library: {
+				type: 'module'
+			},
+			path: path.resolve('templates', 'build')
+		},
+		experiments: {
+			outputModule: true
+		}
 	}
 ];